requirements.txt
a2wsgi==1.10.10
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
apache-airflow==3.0.6
apache-airflow-core==3.0.6
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-common-io==1.6.2
apache-airflow-providers-common-sql==1.28.0
apache-airflow-providers-smtp==2.2.1
apache-airflow-providers-standard==1.7.0
apache-airflow-task-sdk==1.0.6
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
cadwyn==5.4.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.2.0
Flask==3.1.2
fsspec==2025.9.0
googleapis-common-protos==1.70.0
gotrue==1.3.1
greenlet==3.2.4
grpcio==1.75.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.4
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
postgrest==0.10.8
protobuf==6.32.1
psutil==7.1.0
psycopg2==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
PyYAML-ft==8.0.0
realtime==1.0.6
referencing==0.36.2
requests==2.32.5
retryhttp==1.3.3
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
sentry-sdk==2.38.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.48.0
storage3==0.5.4
StrEnum==0.4.15
structlog==25.4.0
supabase==1.0.4
supafunc==0.2.3
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
typer==0.18.0
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.36.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==12.0
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.3
zipp==3.23.0
